Move new MemoryStore classes to Reark library
