+++ app/src/main/java/io/reark/rxgithubapp/advanced/data/stores/StoreBase.java
@@ -30,13 +30,10 @@ import android.support.annotation.NonNull;
 
 import com.google.gson.Gson;
 
-import io.reark.reark.data.store.SingleItemContentProviderStore;
+import io.reark.reark.data.stores.SingleItemContentProviderStore;
 import io.reark.reark.utils.Preconditions;
-import io.reark.rxgithubapp.shared.data.StoreInterface;
-import io.reark.rxgithubapp.shared.data.StorePutInterface;
 
-public abstract class StoreBase<T, U> extends SingleItemContentProviderStore<T, U>
-        implements StoreInterface<U, T> {
+public abstract class StoreBase<T, U> extends SingleItemContentProviderStore<T, U> {
 
     private final Gson gson;
