+++ appbasic/src/main/java/io/reark/rxgithubapp/basic/data/stores/GitHubRepositoryStore.java
@@ -1,7 +1,7 @@
 package io.reark.rxgithubapp.basic.data.stores;
 
+import io.reark.reark.data.stores.MemoryStore;
 import io.reark.reark.utils.Log;
-import io.reark.rxgithubapp.shared.data.StoreInterface;
 import io.reark.rxgithubapp.shared.pojo.GitHubRepository;
 
 /**
