+++ appshared/src/main/java/io/reark/rxgithubapp/shared/network/fetchers/GitHubRepositoryFetcher.java
@@ -32,11 +32,10 @@ import android.support.annotation.NonNull;
 import io.reark.reark.pojo.NetworkRequestStatus;
 import io.reark.reark.utils.Log;
 import io.reark.reark.utils.Preconditions;
-import io.reark.rxgithubapp.shared.data.StorePutInterface;
+import io.reark.reark.data.stores.StorePutInterface;
 import io.reark.rxgithubapp.shared.network.GitHubService;
 import io.reark.rxgithubapp.shared.network.NetworkApi;
 import io.reark.rxgithubapp.shared.pojo.GitHubRepository;
-import io.reark.rxgithubapp.shared.view.RepositoryView;
 import rx.Observable;
 import rx.Subscription;
 import rx.functions.Action1;
