+++ appshared/src/main/java/io/reark/rxgithubapp/shared/network/fetchers/GitHubRepositorySearchFetcher.java
@@ -36,7 +36,7 @@ import java.util.List;
 import io.reark.reark.pojo.NetworkRequestStatus;
 import io.reark.reark.utils.Log;
 import io.reark.reark.utils.Preconditions;
-import io.reark.rxgithubapp.shared.data.StorePutInterface;
+import io.reark.reark.data.stores.StorePutInterface;
 import io.reark.rxgithubapp.shared.network.GitHubService;
 import io.reark.rxgithubapp.shared.network.NetworkApi;
 import io.reark.rxgithubapp.shared.pojo.GitHubRepository;
