+++ reark/src/main/java/io/reark/reark/data/stores/StorePutInterface.java
@@ -1,0 +1,8 @@
+package io.reark.reark.data.stores;
+
+/**
+ * Created by ttuo on 26/06/16.
+ */
+public interface StorePutInterface<T> {
+    void put(T item);
+}
